requirements.txt
flask==1.1.2
slackclient==2.9.3
slackeventsapi==2.2.1
pytest==6.2.1
python-dotenv==0.15.0
flask-sqlalchemy==2.4.4
sqlalchemy==1.3.20